Move error types to wgpu-types
