+++ wgpu/src/api/device.rs
@@ -7,6 +7,7 @@ use crate::api::blas::{Blas, BlasGeometrySizeDescriptors, CreateBlasDescriptor};
 use crate::api::tlas::{CreateTlasDescriptor, Tlas};
 use crate::util::Mutex;
 use crate::*;
+pub use wgt::error::*;
 
 /// Open connection to a graphics and/or compute device.
 ///
@@ -839,89 +840,6 @@ impl From<wgc::instance::RequestDeviceError> for RequestDeviceError {
     fn from(error: wgc::instance::RequestDeviceError) -> Self {
         Self {
             inner: RequestDeviceErrorKind::Core(error),
-        }
-    }
-}
-
-/// The callback of [`Device::on_uncaptured_error()`].
-///
-/// It must be a function with this signature.
-pub trait UncapturedErrorHandler: Fn(Error) + Send + Sync + 'static {}
-impl<T> UncapturedErrorHandler for T where T: Fn(Error) + Send + Sync + 'static {}
-
-/// Kinds of [`Error`]s a [`Device::push_error_scope()`] may be configured to catch.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, PartialOrd)]
-pub enum ErrorFilter {
-    /// Catch only out-of-memory errors.
-    OutOfMemory,
-    /// Catch only validation errors.
-    Validation,
-    /// Catch only internal errors.
-    Internal,
-}
-static_assertions::assert_impl_all!(ErrorFilter: Send, Sync);
-
-/// Lower level source of the error.
-///
-/// `Send + Sync` varies depending on configuration.
-#[cfg(send_sync)]
-#[cfg_attr(docsrs, doc(cfg(all())))]
-pub type ErrorSource = Box<dyn error::Error + Send + Sync + 'static>;
-/// Lower level source of the error.
-///
-/// `Send + Sync` varies depending on configuration.
-#[cfg(not(send_sync))]
-#[cfg_attr(docsrs, doc(cfg(all())))]
-pub type ErrorSource = Box<dyn error::Error + 'static>;
-
-/// Errors resulting from usage of GPU APIs.
-///
-/// By default, errors translate into panics. Depending on the backend and circumstances,
-/// errors may occur synchronously or asynchronously. When errors need to be handled, use
-/// [`Device::push_error_scope()`] or [`Device::on_uncaptured_error()`].
-#[derive(Debug)]
-pub enum Error {
-    /// Out of memory.
-    OutOfMemory {
-        /// Lower level source of the error.
-        source: ErrorSource,
-    },
-    /// Validation error, signifying a bug in code or data provided to `wgpu`.
-    Validation {
-        /// Lower level source of the error.
-        source: ErrorSource,
-        /// Description of the validation error.
-        description: String,
-    },
-    /// Internal error. Used for signalling any failures not explicitly expected by WebGPU.
-    ///
-    /// These could be due to internal implementation or system limits being reached.
-    Internal {
-        /// Lower level source of the error.
-        source: ErrorSource,
-        /// Description of the internal GPU error.
-        description: String,
-    },
-}
-#[cfg(send_sync)]
-static_assertions::assert_impl_all!(Error: Send, Sync);
-
-impl error::Error for Error {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match self {
-            Error::OutOfMemory { source } => Some(source.as_ref()),
-            Error::Validation { source, .. } => Some(source.as_ref()),
-            Error::Internal { source, .. } => Some(source.as_ref()),
-        }
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Error::OutOfMemory { .. } => f.write_str("Out of Memory"),
-            Error::Validation { description, .. } => f.write_str(description),
-            Error::Internal { description, .. } => f.write_str(description),
         }
     }
 }
